requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cleanup==5.1.0
django-crispy-forms==1.11.1
django-smartfields==1.1.3
django-widget-tweaks==1.4.8
Pillow==8.1.0
pytz==2021.1
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
